ai-service/requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-mistralai>=0.2.0
langchain-core>=0.3.0

# FastAPI and API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Vector Database
qdrant-client>=1.7.0

# Utility dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Optional but useful
typing-extensions>=4.8.0

# Text-to-Speech
elevenlabs>=1.0.0

# Tracing
langfuse>=2.0.0

